Frontend search page (#2)
Ready to be reviewed. #3 was forked from the tip of this branch so I'll
need to rebase it after this is merged. Thus any review comments here,
to avoid merge confusions, I'd like to incorporate into the next PR,
including making the pagination params and such match what the backend
will have by default from Django.