Frontend cart, finish MVP (#3)
Ready to review. This basically finishes implementing all of the
functionality detailed in [the
wireframe](https://github.com/user-attachments/files/23199523/meta2onto_wireframes.html).
It doesn't attempt to tackle all of the features talked about in the
user stories document, such as thumbs upping highlights to improve
search relevance. But it should serve as a good MVP. It should be very
ready to hook up to the backend once it is ready.